Changed the setAlpha function in TransparencyInterface to be a private method. You can now adjust the alpha of the current transparency state directly by using the CurrentAlpha property.
Got rid of the concept of a "current" layer state from the LayerController. This no longer makes sense as the layer states are not always the same for each window. The only thing that was using the current state was the LayerGUIController to synchronize its menus. This has been modified to look at the transparency objects directly instead of the layer state.

Removed the applyLayerStateTemporaryUndisruptive function. The ImageRenderer now creates its own transparencystate and uses this to control what is shown when it is rendering.

git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@1212 956a8c8d-885e-cb46-a0b9-a2f355b4c364